+++ Panther/include/Context.h
@@ -353,9 +353,10 @@ namespace pcit::panther{
 			auto add_task_to_work_manager(SymbolProc::ID symbol_proc_id) -> void {
 				if(this->hasHitFailCondition()){ return; }
 
+				const SymbolProc& symbol_proc = this->symbol_proc_manager.getSymbolProc(symbol_proc_id);
+
 				evo::debugAssert(
-					this->symbol_proc_manager.getSymbolProc(symbol_proc_id).status == SymbolProc::Status::IN_QUEUE,
-					"Invalid status to add to work manager"
+					symbol_proc.status == SymbolProc::Status::IN_QUEUE, "Invalid status to add to work manager"
 				);
 
 				this->work_manager.visit([&](auto& work_manager) -> void {
@@ -365,10 +366,18 @@ namespace pcit::panther{
 						evo::debugFatalBreak("Cannot add task to work manager as none is running");
 
 					}else if constexpr(std::is_same<WorkManager, core::ThreadQueue<Task>>()){
-						work_manager.addTask(symbol_proc_id);
+						if(symbol_proc.isPriority()){
+							work_manager.addPriorityTask(symbol_proc_id);
+						}else{
+							work_manager.addTask(symbol_proc_id);
+						}
 
 					}else if constexpr(std::is_same<WorkManager, core::SingleThreadedWorkQueue<Task>>()){
-						work_manager.addTask(symbol_proc_id);
+						if(symbol_proc.isPriority()){
+							work_manager.addPriorityTask(symbol_proc_id);
+						}else{
+							work_manager.addTask(symbol_proc_id);
+						}
 
 					}else{
 						static_assert(false, "Unsupported work manager");